create migration for admin users

+++ database/migrations/2021_03_30_040248_create_enrollment_regiment_form.php
@@ -13,24 +13,26 @@ class CreateEnrollmentRegimentForm extends Migration
      */
     public function up()
     {
-        Schema::create('enrollment_regiment_form', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('form_id');
-            $table->foreign('form_id')->references('id')->on('tb_mac_forms')->onDelete('cascade');
-            
-            $table->string('treatment_history')->nullable();
-            $table->string('registration_group')->nullable();
-            $table->string('risk_factor')->nullable();
-            $table->string('drug_susceptibility')->nullable();
-            $table->string('current_weight')->nullable();
-            $table->string('suggested_regimen')->nullable();
-            $table->string('regimen_notes')->nullable();
-            $table->string('clinical_status')->nullable();
-            $table->string('signs_and_symptoms')->nullable();
-            $table->string('vital_signs')->nullable();
-            $table->string('diag_and_lab_findings')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('enrollment_regiment_form')) {
+            Schema::create('enrollment_regiment_form', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('form_id');
+                $table->foreign('form_id')->references('id')->on('tb_mac_forms')->onDelete('cascade');
+                
+                $table->string('treatment_history')->nullable();
+                $table->string('registration_group')->nullable();
+                $table->string('risk_factor')->nullable();
+                $table->string('drug_susceptibility')->nullable();
+                $table->string('current_weight')->nullable();
+                $table->string('suggested_regimen')->nullable();
+                $table->string('regimen_notes')->nullable();
+                $table->string('clinical_status')->nullable();
+                $table->string('signs_and_symptoms')->nullable();
+                $table->string('vital_signs')->nullable();
+                $table->string('diag_and_lab_findings')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
